agents/a2a_agents/hitl_agent/pyproject.toml
[project]
name = "roundrobin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [ "a2a-sdk==0.3.0",
    "autogen-agentchat>=0.7.4",
    "autogen-ext[openai]>=0.7.4",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",]
